FINALLY FIXED THE SAVE MECHANISM!
Holy shit this took hours of debugging to see where the problem EXACTLY
are.
In short the number that was returning was still one EVEN after going
into the index checker.  Also with further checking it seemed like the
number array was resetting.

Woo.... Hopefully we have never change things in here anymore..