Tons of changes in the discription!
Re-enable disable buttons on transition / popup.

Expanded the colouring of the buttons to: Oefentoets, Operator en Toets.
Fixed the interaction between all of them. Even fix the button: "Opnieuw
Beginnen". It now removes all the colours within it's operator.

Added unset whenever a student scores lower than 5.5 in the "Oefentoets"
and "Toets". It unset assignments for the "Oefentoets" and "Oefentoets"
for the "Toets".

Added restriction to the "Toets". You need to finish all "Oefentoets"
before you can do the "Toets".